// cgocore-ui/src/app/features/models/loading/views/full/DatedFlightUldAwbFullView.ts
import { DatedFlightUldAwbForm } from 'src/app/features/models/loading/forms/DatedFlightUldAwbForm';
/**
 * auto generated full veiw bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class DatedFlightUldAwbFullView {

id: number;
form: DatedFlightUldAwbForm;
canUpdate: boolean;
canDelete: boolean;

/* Specific Code Start */
/* Specific Code End */
}